오버플로우 해결

+++ chap03/src/chap03/OverflowExample.java
@@ -3,9 +3,9 @@ package chap03;
 public class OverflowExample {
 
 	public static void main(String[] args) {
-		int x = 1000000;
-		int y = 1000000;
-		int z = x * y;
+		long x = 1000000;
+		long y = 1000000;
+		long z = x * y;
 		System.out.println( z );
 	}
